fix: address all defects found by subagent deep-scan
protocol/jsonrpc.rs:
  - Add PartialEq + Eq to JsonRpcError/Request/Notification/Response
  - Fix deser priority: check Response before Request (JSON-RPC spec)
  - Replace unwrap_or_default (produces Value::Null) with {} fallback
  - Fix comment: UUID version nibble at position 14, not 12

protocol/handshake.rs:
  - Rename PROTOCOL_VERSION (shadowed lib.rs &str) to HANDSHAKE_VERSION

ws/services/llm_provider.rs:
  - Add #[ts(rename = type)] on entry_type (matches serde rename)

ws/services/knowledge_base.rs:
  - Add #[serde(default)] #[ts(optional)] on knowledge_base field

http.rs:
  - Add #[serde(default)] on SceneCamera::bookmarks (was missing)
  - Add #[serde(default)] on all 4 UserPreferences fields (deser fails on omit)

mcp/aporia.rs + philia.rs + haplotes.rs:
  - Add #[ts(export)] on AnomalyResult (aporia), 12 structs (philia), and
    LlmProviderCallParams/SubscribeTriggerParams (haplotes)
  - Remove stale planning comment from haplotes.rs
  - Keep ConversationContext with doc comment (semantically distinct)

mcp/hubris/skemma/polemos:
  - Fix type mismatches (i64→i32, u64→u16) between params and results

mcp/neikos.rs:
  - Rename DeliverMessageResult -> SidecarDeliverResult (collision with
    epieikeia's identically-named but different type)

ws/ui/views.rs + workspace.rs + noa.rs:
  - ViewKind: add #[serde(rename_all = snake_case)]
  - NoaEvent.metadata: align ts type to Record<string, unknown>
  - connection_kind: add #[ts(optional)]

enums.rs + tracing_helpers.rs:
  - str_enum! macro: add Hash + JsonSchema derives
  - ShortTimer: add Debug + Clone + Copy + Default derives

+++ src/mcp/hubris.rs
@@ -83,7 +83,7 @@ pub struct ClearTodoParams {
 pub struct MoveTodoParams {
     pub todo_id: Uuid,
     pub new_parent_id: Option<Uuid>,
-    pub sort_order: Option<i64>,
+    pub sort_order: Option<i32>,
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize, ts_rs::TS)]
